now fix building with non-MinGW GCC

+++ Source/net_client.c
@@ -770,7 +770,7 @@ static void NET_CL_ParseGameData(net_packet_t *packet)
 
         if (!NET_ReadFullTiccmd(packet, &cmd, settings.lowres_turn))
         {
-            NET_Log("client: error: failed to read ticcmd %u", (unsigned int)i);
+            NET_Log("client: error: failed to read ticcmd %lu", (unsigned long)i);
             return;
         }
 
@@ -787,7 +787,7 @@ static void NET_CL_ParseGameData(net_packet_t *packet)
 
         recvobj->active = true;
         recvobj->cmd = cmd;
-        NET_Log("client: stored tic %u in receive window", (unsigned int)seq + i);
+        NET_Log("client: stored tic %lu in receive window", (unsigned long)seq + i);
 
         // If a packet is lost or arrives out of order, we might get
         // the tic in the next packet instead (because of extratic).
